Moved into correct files, and added payment example
